requirements.txt
fastai==1.0.60
torch==1.5.0
torchvision==0.6.0
Pillow==7.1.1
Flask==1.1.1
requests==2.22.0
gunicorn==19.9.0
